Use uint IDs; add encoder & playback worker
Convert Track ID and Duration from int to uint across models, services, handlers and repositories for consistent unsigned identifiers. Update repository interfaces and SQLite implementations (including LastInsertId assignment) to use uint and add a GetByID implementation. Modify scheduler: localize error messages to Russian, change RegisterStation to unlock before refilling and ignore ErrNoTracks. Introduce a transcoder encoder (TrackStreamer/Encoder) that streams audio via ffmpeg, add a playback worker (runStationWorker) to pull and stream tracks, and update PlaybackEngine to accept a TrackRepository, scheduler and TrackStreamer; NewStation now spawns the station worker.

+++ internal/api/handlers/trackHandler.go
@@ -179,7 +179,7 @@ func (h *TrackHandler) createTrackFromJSON(w http.ResponseWriter, r *http.Reques
 		Artist:   payload.Artist,
 		Album:    payload.Album,
 		Path:     payload.Path,
-		Duration: payload.Duration,
+		Duration: uint(payload.Duration),
 	}
 
 	if err := h.trackService.Create(track); err != nil {
@@ -204,11 +204,11 @@ func (h *TrackHandler) UpdateTrack(w http.ResponseWriter, r *http.Request) {
 	}
 
 	track := &models.Track{
-		ID:       id,
+		ID:       uint(id),
 		Title:    payload.Title,
 		Artist:   payload.Artist,
 		Album:    payload.Album,
-		Duration: payload.Duration,
+		Duration: uint(payload.Duration),
 	}
 
 	err = h.trackService.Update(track)
